# Improve performance of nvimcom launch