@jalvesaq 2023-11-12
I decreased the number of times that `installed.packages()` is called in `before_nrs.R`. The function is still called once because in my system (Linux) calling `find.package()` multiple times is slower than calling `install.packages()` once. The time of running `before_nrs.R` decreased only from about 0.20 to about 0.18 of a second. Now, you could:

- Either try to further improve the time of running `before_nrs.R` or remove it from your pull request.
- Increase `nvimcom` version number at `R/nvimcom/DESCRIPTION`.
- Do the pull request.