@klmr 2023-11-13
0.2s vs 0.18s is a tiny difference: I don’t think it makes sense to optimise for a few millisecond.

The difference I am observing on my system is an order of magnitude larger: even calling `install.packages()` *just once* can take more than 5 seconds on a cold start (when there’s on OS level filesystem caching) and always takes >4s. Conversely, calling `find.package()` repeatedly for every package in `getOptions("defaultPackages")` take a few milliseconds.

So removing `install.packages()` would be beneficial: in the worst case (= fast local storage, few packages installed) it makes the launch *imperceptibly* slower. But in the best case (= slow storage, many packages installed) it makes it several seconds faster.