@klmr 2023-11-13
Here’s a fairer benchmark which calls both `find.package()` *and* `packageDescription()` on all default packages:

```r
getOption('defaultPackages')
```

> ```
> # [1] "datasets"  "utils"     "grDevices" "graphics"  "stats"     "methods"
> ```

```r
bench::mark(
  `installed.packages()` = callr::r(\() installed.packages()[c('abind', 'zoo', 'nvimcom'), ]),
  `packageDescription()` = callr::r(\() lapply(getOption("defaultPackages"), \(p) if (length(find.package(p, quiet = TRUE)) == 1L) packageDescription(p))),
  iterations = 5L,
  check = FALSE,
  memory = FALSE
)
```

> ```
>   expression             median `itr/sec` total_time
> 1 installed.packages()    1.48s     0.440      9.09s
> 2 packageDescription()  517.8ms     1.92       2.61s
> ```

And the median `installed.packages()` time is very misleading (see the `total_time`): when I use `Rscript -e` from the command line instead of `callr::r()` from inside R, I reproducibly get times > 4s the first time, and lower times afterwards (and I think this is fairly representative for how FS-level caching works):

```r
cmd = file.path(R.home('bin'), 'Rscript')
args = c('-e', shQuote('x <- installed.packages()'))

for (i in 1 : 5) print(system.time(system2(cmd, args))[['elapsed']])
```

> ```
> [1] 4.89
> [1] 1.834
> [1] 1.671
> [1] 2.614
> [1] 1.793
> ```