@jalvesaq 2023-11-13
Indeed, I rebooted my laptop and the time to run `before_nrs.R` was 5 seconds.