@jalvesaq 2023-11-13
Now I see why your pull request was a draft. It impacted other parts of Nvim-R code. I implemented your idea of not calling `installed.packages()`, making all necessary adjustments in other parts of Nvim-R. Thanks for your contribution, and feel free to optimize the code further.