@jalvesaq 2023-11-13
You can close the PR, unless you still want to work on the issue. I will fix the missing `quiet=TRUE` and `verbose=FALSE` that you noted.